+++ src/main/java/com/github/alexeysol/geekregimeapiposts/repositories/PostRepository.java
@@ -1,6 +1,6 @@
 package com.github.alexeysol.geekregimeapiposts.repositories;
 
-import com.github.alexeysol.geekregimeapiposts.entities.Post;
+import com.github.alexeysol.geekregimeapiposts.models.entities.Post;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
     @Query("SELECT p FROM Post p WHERE p.slug = :slug")
     @Transactional
     Post findPostBySlug(String slug);
@@ -16,7 +16,7 @@ public interface PostRepository extends CrudRepository<Post, Integer> {
     @Query("DELETE FROM Post p WHERE p.id = :id")
     @Transactional
     @Modifying
-    int removePostById(int id);
+    int removePostById(long id);
 
     Boolean existsPostBySlug(String slug);
 }
